requirements.txt
aiogram==2.17.1
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.0.6
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
bcrypt==3.2.0
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.10
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cleanup==5.2.0
django-filter==21.1
django-modeltranslation==0.17.3
django-rosetta==0.9.8
django-timezone-field==4.2.1
frozenlist==1.2.0
idna==3.3
kombu==5.2.2
multidict==5.2.0
Pillow==8.4.0
polib==1.1.1
prompt-toolkit==3.0.24
psycopg2==2.9.2
pycparser==2.21
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-slugify==5.0.2
pytils==0.3
pytz==2021.3
requests==2.26.0
six==1.16.0
slugify==0.0.1
sqlparse==0.4.2
text-unidecode==1.3
typing_extensions==4.0.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
yarl==1.7.2
